Some fixes

+++ mainForm.cs
@@ -8,6 +8,7 @@ using System.Text;
 using System.Threading.Tasks;
 using System.Windows.Forms;
 using System.IO;
+using System.Diagnostics;
 
 namespace AutoInitio
 {
@@ -25,9 +26,23 @@ namespace AutoInitio
             this.WindowState = FormWindowState.Normal;
         }
 
-        private void buttonScheduler_Click(object sender, EventArgs e)
+        private void buttonScheduler_Click(object sender, EventArgs e) // ToolTip
         {
             panelMain.Controls.Clear();
+
+            Process process = new Process();
+            process.StartInfo.FileName = "schtasks.exe";
+            process.StartInfo.Arguments = "/query /fo csv /nh";
+            process.StartInfo.UseShellExecute = false;
+            process.StartInfo.RedirectStandardOutput = true;
+            process.Start();
+            StreamReader reader = process.StandardOutput;
+            string[] output = reader.ReadToEnd().Split('\n');
+            foreach (string str in output)
+            {
+                textBox1.Text += str + '\n';
+            }
+            process.WaitForExit();
         }
 
         private void buttonRegistry_Click(object sender, EventArgs e)
@@ -39,7 +54,7 @@ namespace AutoInitio
         {
             string[] files = Directory.GetFiles(Environment.GetFolderPath(Environment.SpecialFolder.Startup));
 
-            if (files.Length == 0 || Path.GetFileName(files[0]) == "desktop.ini")
+            if ((files.Length == 1 && Path.GetFileName(files[0]) == "desktop.ini") || files.Length == 0)
             {
                 Label label = new Label();
                 label.Location = new Point(0, 0);
@@ -58,8 +73,7 @@ namespace AutoInitio
                         nodesFolders.Add(file);
                     }
                 }
-
-                updatePanel();
+                updateFolders();
             }
         }
 
@@ -69,14 +83,20 @@ namespace AutoInitio
 
             if (button != null)
             {
-                MessageBox.Show("Вы уверены, что хотите удалить " + Path.GetFileName(nodesFolders[button.index]) + " ?", "Подтверждение", MessageBoxButtons.YesNo, MessageBoxIcon.Question);
+                var result = MessageBox.Show("Вы уверены, что хотите удалить " + Path.GetFileName(nodesFolders[button.index]) + " ?", "Подтверждение", MessageBoxButtons.YesNo, MessageBoxIcon.Question);
+                if (result == DialogResult.No)
+                {
+                    // cancel the closure of the form.
+                    //e.Cancel = true;
+                    return;
+                }
                 File.Delete(nodesFolders[button.index]);
                 nodesFolders.Remove(nodesFolders[button.index]);
-                updatePanel();
+                updateFolders();
             }
         }
 
-        private void updatePanel()
+        private void updateFolders()
         {
             panelMain.Controls.Clear();
 
@@ -84,12 +104,13 @@ namespace AutoInitio
             {
                 Label label = new Label();
                 label.Location = new Point(10, 15 + (i * 30));
+                label.AutoSize = true;
                 label.Text = Path.GetFileName(nodesFolders[i]);
 
                 myButton button = new myButton();
-                button.Location = new Point(120, 10 + (i * 30));
+                button.Location = new Point(410, 10 + (i * 30));
                 button.index = i;
-                button.Text = "кнопка" + button.index;
+                button.Text = "Удалить";
                 button.Click += button_Click;
 
                 panelMain.Controls.Add(label);
